Add files via upload
This is an attendance management for BE graduation program under computer engineering branch for semester 3 to 8. This project saves all the data inside inside a .xlsx file using pandas' xlsxwriter module, that includes the only pre-requisite, that is the file and for that purpose, have upload sample .xlsx file that is just for the sake of file name. If file name is not same the system doesn't works, else it requires to change the location in the python code.